로그인 (15) - reissue 기능 구현 2

+++ src/main/java/com/my/bob/entity/BobUser.java
@@ -22,7 +22,7 @@ public class BobUser extends BaseTimeEntity{
     @Id
     @Column(name = "USER_ID")
     @GeneratedValue(strategy = GenerationType.IDENTITY) // DB에게 ID 생성을 맡김
-    private long userId;
+    private int userId;
 
     @Column(nullable = false)
     private String password;
